Move KeyVaultPropertyInitializer to KeyVaultPropertyEnvironmentPostProcessor
See gh-162
